Create new Class Ui. Shift method readUserInput to Class Ui
