+++ data/output.txt
@@ -3,3 +3,5 @@ T | 1 | shit
 T | 0 | eat
 T | 1 | olist
 T | 0 | go toilet
+T | 1 | go delta
+E | 1 | coaching | 6pm | 7pm
